mass commit of existing tools
This just puts all the encoding scripts and artifacts
into the repo initially. Many of these things were
developed years apart. Some depend on others and may
need paths adjusted to continue working.

This is just a first stab at getting all this version
controlled because the drive they're on has been in
use for a looooong time. It would be a pity if it
failed and I had to rewrite all this from scratch.