start of intelligent movie encode script
I'm copying the BSG smart-encode.sh script and trying to adapt it to
process arbitrary movie directories in a similar way, relying on
manual config files for the parts that can't be guessed
automatically. In BSG, the file names were unique because I'd already
processed them some. For this, I think the file name and parent
directory (which should be named for the movie) will have to make up
the "uniqueness" for any given file.

This is in no way meant to work right now. It's just a baseline to
start from.