done.sh supports multi-level paths for log files
For processing movies, each movie will be in a separate directory for
input, and outputs will sometimes be in the nested plex directory
structure. To support this, the done script needs to be able to
understand file names that aren't just in a flat directory. This keeps
the original behavior for backward compatibility and adds the ability
to pass a relative path for using for the log lookup.