interlaced.sh handles ambiguous progressive movies
Some movies show lots of undetermined frames on the single or multi
line, and this was flagging them as undetermined. In the cases I've
seen, the number of TFF/BFF frames is usually nearly zero, and the
number of progressive frames is usually somewhat high. This adjusts
the script to identify those as progressive.