adding a bunch of configs
This is the first big group of configs that tell the tools how to rip
things. Maybe I can figure a more intelligent way to detect patterns
in the media streams, but for now, I manually create them all, so
storing them in version control is a good safety measure.