way better cropping detection
I sort of forgot my crop detection was very experimental and then
realized it wasn't doing a very good job. In retrospect, how could it
by examining a random 10-second segment of a video, especially only 5
minutes in, when some movies are still in opening credits?

This overhauls the crop.sh script to be much more comprehensive while
not being slow as molasses. If you just scan the whole movie, it takes
a considerable amount of time. Even sampling every 10th or 100th frame
doesn't improve things much. The time seems to be in streaming the
whole video. You can speed it up by taking multiple small chunk with
-ss and -t. Those are very fast. This looks at the overall length,
slices it into chunks, and takes a 2-second sample every so often,
shooting for 120 samples evenly spaced through the movie, ignoring 10%
at the start and end for opening and closing scenes.

It picks the largest value found by looking at the max height and
width found. If the width doesn't match the video width, it
aborts. That will probably need to be fixed because there are lots of
fullscreen specials.

Also removes all CROPPING entries from configs because I hope this
will make it all a lot more automatic.