validate script allows different cuts
Previously the script required exactly one movie file in a movie dir
and required it to be named the same as the movie.  This was to
prevent typos on my part. It doesn't work when I have multiple cuts of
a movie because there will be two or more files.  They should still
have the same initial name, though, so I should be able to match with
a regex.