big overhaul of crop.sh and interlaced.sh
While trying to tweak interlaced.sh so I wouldn't have to manually
configure it so often, I realized it might be having a similar problem
to crop.sh, which is that it's looking at a very small sample (1000
frames) of the video.

I started adapting it to take a similar approach to crop.sh and sample
the input video, and then I realized that both cropping and interlace
detection are just ffmpeg filters and could be run in the same ffmpeg
command.

That led me to this approach, which now runs a single analysis pass
using analyze.sh. It samples the input video, applies approriate
filters, and writes the output to file, then calls crop.sh and
interlaced.sh to inspect the output instead of each of them reading
the video file themselves.