several accumulated updates
This isn't a nice commit with a single change. It's changes that have
been made over the course of weeks/months. Mainly, as I recall, I
added the ability to make specific configs to BSG files and improved
some of the local discovery scripts. There are also several smaller
changes to other pieces of this encoding thing.