fixes encode.sh video filtering
Like I thought in a recent commit, the script is guaranteed to fail if
a complex filter gets built and either cropping or deinterlacing is
needed because both the simple and complex filter strings will be
set. This fixes that by only setting one or the other, depending on
whether cutting is needed.