add ability to specify cache file locations
Some helper scripts are able to cache results to speed up later
invocations. Results are cached in files, and the location of the
cache file can vary based on the context in which they're
invoked. This adds an option to the script to let callers specify the
location of the cache file. If not specified, they fall back to a
default location derived from the input path. The default location
isn't very useful except during debugging.