add next iteration of encoding automation
This creates a new script that draws on previous lessons learned. I'm
trying to make an encoding script that can deal with both movies and
tv shows, supporting all the features I've previously invented for
both, includeing at least: trimming objectionable content from any
given media file, recognizing and transcoding tv episodes according to
a pre-known "profile", and all of the other features previously
available to either tv or movie encoding in this project (I might be
missing some at the moment). This started as a copy of encode.sh from
the 'movies' dir, and it's expanded to support tv shows.