make locations configurable
Previously, I've always just switched to the directory containing
`encode.sh` to run it, in this and previous iterations of the
script. There are already variables for most of the important
locations at the top of the file. This just makes them configurable
and makes the script fail if any of the locations doesn't exist.
This lets me easily grab this project and run it from a new location
where things exist at different paths.