add House episodes, first pass
This is automatically assigned episode numbers based on the file order.
When ripping, all the episodes are in order, and you can tell episodes
from extras by file size. I think these numbers are all correct, but
there's an episode missing from season 2 disk 5 because of ripping
errors. The disk is getting unreadable. I'll try to get the episode
again.