add nvidia gpu encoding
This increases transcoding time by 10x! It also massively improves the
throughput of the interlaced.sh script. The analyze.sh and crop.sh
scripts don't work, possibly because of the piping. Try reorganizing the
code so that it's just a single ffmpeg command with no piping. Maybe
they'll get the same speed benefits.