make sampling support writing to file
In order to make further use of gpu encoding, I looked at the analyze.sh script, which is currently breaking. I
suspect it's because of the ffmpeg piping it's using. This modifies sample.sh to be able to either write to a pipe
or to a file so that the analyze script can have it write to a temp file and then read that for analysis. It's
probably less efficient this way, but hopefully the time saved by using gpu encoding will more than make up for it.

A couple other scripts also use sample.sh under some conditions, but I don't think those conditions are actively
used right now, so I just put a time bomb in it that I'll hit later if I try to use them that way again.