make filter.sh work with configurable dirs
Since making the encode dirs configurable a few commits back, I hadn't
run into anything that used filter.sh until now. That script was still
assuming a local data dir. This updates it to match encode.sh, but I
think it would be better to have the dirs passed to it so that this
isn't duplicated across files. I'll just have the same problem the next
time I update this stuff.