add new scripts for managing tv show config
This adds a script similar to newmovie.sh to add new tv show disks. There's also a companion
script for generating thumbnails of all the titles in a disk folder so you can scan over them
and find the garbage titles, like main menu backgrounds. It works moderately well, but it
would be better if it did a better job of sampling titles based on their length.