add script to find title overlap
One of the hardest things in ripping stuff is going through all the titles of a disk to figure
out what's what. One of the most annoying things is when you find title that overlap with others.
Sometimes it's an exact duplicate without chapter marks. Sometimes several titles are parts of
another, aggregate title. This adds a script to help find those overlaps before the disk is
even ripped. It can look at makemkvcon info output and compare title segments to help understand
how titles are interrelated.