Revert "encode.sh omits non-english streams"
This reverts commit 383942242f06dfb2d17a62a6a87af905ac0a36d8.

Instead of relying on this, I found all the rips that contained non-english streams
and fixed them. The offenders consisted of some BSG and Pitch Black titles that
had "spa" and "fra" subtitles. I copied all the other streams out to a temp file
and then replaced the original with that file.