make check-overlap-title script work a little better
The regex ending in $ works fine in ubuntu, but git bash on windows doesn't seem to like it.
Removing it should be fine. I just prefer to be explicit when matching stuff.

Also added a line of output that confirms what was found because when I first tried this,
it ran and just produced no output.