add manual interlace settings to a few titles
Automatic detection wasn't working for these. I've noticed that many of the titles it fails on are
special features that combine clips from a movie with other stuff, like interviews. If the movie
is interlaced and the interviews aren't, then obviously it's not going to be clear whether to
deinterlace or not. I'm thinking I should adjust the interlace detection to be more likely to
detect a video as "interlaced".