make gpu encoding work with software filters
I was getting errors on some titles about not being able to find a compatible
frame format between two filters, or something like that. Some searching around
suggested this was because I have the stream in the GPU hardware, and it doesn't
support the filters being used (like yadif, for deinterlacing). The solution is
to copy the frames back from the gpu, run the filters, and then copy them back
to the gpu, which is what the hwdownload and hwupload filters do. Apparently
something here is really unintelligent because you also have to set the pixel
format. It's the first time I've had to deal with that, but from what I can tell,
nv12 is a pretty ubiquitous format.