cache at top level only
The caching stuff has gotten annoying to manage since moving to remote cache dirs.
This adds caching at the top-level encode script and turns it off in the lower-level
scripts. It's still there as an option, but off by default. That's except for crop.sh,
which had some caching logic but didn't write out cache files even when caching was
enabled.

This makes caching granularity coarser, so that if part of an analysis fails, the whole
thing will have to run again, but that doesn't happen anymore. It's all pretty stable.
I think this will have an overall positive impact on the efficiency of ripping things
and advancing the project.