fix interlaced.sh for new sample inputs
While working out how to do cuda encoding, I found that the way sampling
was happening didn't work, and I tried fixing it. It turns out he pipe
wasn't the only problem, so this just leaves it there but using the new
input options supported by sample.sh. As I learn more about GPU encoding,
it's possibly I might be able to make this use it, too. It's the slow
part of the whole process now.