newmovie.sh supports standalone features
How to handle shorts that have their own movie DB entries, such as the Marvel One Shots?
I want them to show up as separate entries in the movie library, so I don't have to go
hunting through all marvel movies to find them. It took a bit of thinking, but I realized
everything is already in place to manage it. Since config is inherited, all I have to do
is overrided the movie MAIN_NAME in a title config file, and it just works. The title still
"belongs" to a disk, but it gets encoded as a separate movie.