newmovie.sh doesn't require name
Requiring name and directory to be passed into this script makes it more
cumbersome to use from one movie to the next: a lot of stuff to delete and
retype in the command. If I require only the directory, I can make a guess
at the movie name, which will be right sometimes and give me something to
copy/paste sometimes. Overall, it reduces the amount of command rewriting
necessary.