make title overlap checker more robust
Many times I'd run this script and pass only the movie dir, not the full path to
the _info file. It occurred to me that there's no reason that shouldn't work. This
makes it so if you pass a dir to the script, it looks for an _info file inside it.
You can pass either the directory or the file.