remove bogus home alone 2 config
I renamed the config file for some reason, and it got committed by accident,
I think. There's no need for this file, and it's a duplicate of the correctly
named one. I probably moved it out of the way so I could encode the extras
without waiting for the main movie?