add more (final?) star trek stuff
This is the last few disks of episodes, in addition to configs
for all the specials, though I haven't explicitly mapped any.
I just found and excluded a bunch of titles by listing them
all out sorted by length. A lot of the repetitive stuff is all
the same length, like: tv teasers of next episode, backgrounds
for blu-ray menus, and some other things that are repeated across
multiple disks.