add first two matlock seasons
These configs are all auto-generated except for the very first one. I'm assuming,
based on some spot checks, that the episodes are all in order on the disks. I'll
check further after getting everything ripped. I'm committing this so I'll notice
if future runs of my command alter anything that was written before.