account for counts of 0 in interlaced.sh
As mentioned in the comment, ffmpeg on ubuntu 23.10 is emitting data like

```
[Parsed_idet_0 @ 0x5619e187a900] Repeated Fields: Neither:     0 Top:     0 Bottom:     0
[Parsed_idet_0 @ 0x5619e187a900] Single frame detection: TFF:     0 BFF:     0 Progressive:     0 Undetermined:     0
[Parsed_idet_0 @ 0x5619e187a900] Multi frame detection: TFF:     0 BFF:     0 Progressive:     0 Undetermined:     0
[Parsed_idet_0 @ 0x5619e189ddc0] Repeated Fields: Neither: 68597 Top:  8469 Bottom:  7629
[Parsed_idet_0 @ 0x5619e189ddc0] Single frame detection: TFF: 64323 BFF:    26 Progressive: 12218 Undetermined:  8128
[Parsed_idet_0 @ 0x5619e189ddc0] Multi frame detection: TFF: 84570 BFF:    23 Progressive:    73 Undetermined:    29
```

The regex matches the first matching lines, and the script comes up with
zero for all the counts, which ends up making it decide everything is
progressive. I added checks that will fail if this is the case in the
future and also a grep filter that should remove all of these bogus lines
from the output, letting the script match the real lines.