add a script for checking cached analyses
The interlaced script has been giving bad results (see recent commit).
That means lots of the cached analysis results are wrong. Rather than
nuking the whole dir and re-encoding everything, I'll see if I can
find titles whose interlaced setting don't match what the script
currently says it is. This probably isn't a useful long-term thing,
but it'll take quite a while to run, so I want to be able to shift
it between different locations and update the file incrementally.
The script is ugly because I was just running it at the CLI until
I realized I'd want to move it around.