add interlaced setting to martin luther
I'm seeing that the interlaced script will call something interlaced
if the interlaced frames are ~300 and the progressive frames are ~1,
for example, even if undetermined is in the thousands. This is because
of the "greatly outweighs" clause. Maybe I should fix it? It's making
the right call here, though, but when sampled, the ~300 becomes
something less than 100, making the the script fail to come to an
answer.