add BSG cut files from before
This translates the cuts files from the bsg dir to the allinone dir.
It wasn't straightforward because last time, I had based the cuts
on file names I'd already translated to season+episode, so I scripted
up an automatic copy from that to the new structure. I had to work
backward from the output name calculated by the encode script and
match up the old cut file to a source file that way.