account for warning messages and thumbnails in filter.sh
First, this script was failing on a probesize warning, which had
something like "Stream #0:6" in it, triggering the stream match.
Increasing probesize fixed that. Then, it failed on a thumbnail
stream as follows:

```
+ [[ Stream #0:6: Video: mjpeg (Baseline), yuvj444p(pc, bt470bg/unknown/unknown), 640x360 [SAR 72:72 DAR 16:9], 90k tbr, 90k tbn (attached pic) =~ Stream #(.:.)\([^)]*\): (Video|Audio|Subtitle): ]]
+ echo 'Couldn'\''t detect stream type'
Couldn't detect stream type
+ exit 1
```

I added a clause to make it ignore streams that look like this.