account for cutting with multiple streams
When I fixed up cutting to work on subtitles, I decided to get rid of the
CUT_STREAMS setting and instead keep all the streams that would be kept even
if not doing any cutting. The encode logic was only built to handle a single
video and a single audio stream, so now when cutting a video, it tries to do
a streamcopy on any other streams, which isn't compatible with a complex
filtergraph. This makes it generate audio encoding args for each audio stream
being transcoded instead of being hardcoded for a single audio stream.