encode script supports tv specials as movie
I previously made the newtvshow script able to create config files that cause tv specials
to be stored as movies, but for existing things, I need the encode script to be able to do
it. House specials are an example. I had them set with an output name to be stored as movies,
but since the input paths contained "Season", the script was putting them in the tv output dir.
This just adds a new config variable that the script will check to tell where to put tv specials.
The logic for determining if something is a special might be a little flimsy, but it's working
on the House specials.