add new script for controlling input listings
For a long time, I've wanted better control over how the encode script
gets its list of files. This introduces a new script dedicated to
listing input files. Following the Unix philosophy, it does one thing
and does it well, I think. I should be able to take this and xargs it
into the encode script as is. If all goes well, the file listing stuff
can be taken out of encode, and it can be a flat script instead of a big
function wrapped in a script.