add Full House configs
Again, a show that seems to have titles in order. I did easily with
minor modifications to the script I used for Walker in the previous
commit. This one was simpler. I'm thinking there's a reusable
"create-tv-configs" script in here that could handle all kinds of
variants via cli args.