fix flintstones s1d3 mappings
When doing more scripted tv show mapping, I stumbled across an issue
here when I ran my script and compared it with the existing configs.
Somehow, the rip of this disk seems to have been corrupted. Note the
inconsistent file names. When I checked the rip dir, it turns out that
the titles with shorter names were duplicates of the other titles, and a
few episodes were missing! I re-ripped the disk, and it came out fine
this time. This updates the config files to match, and I ensured they're
correct this time.