add muppet show configs
I used the new "auto rip tv shows" script, which worked great. The
problem here is that there are multiple different orders for this
series. The tvdb dvd order is correct here. I'm not sure if I can make
Plex match that, since it moved to tmdb for tv shows.