add seasons 1 and 2 of Transformers with caveat...
This DVD series is troublesome. Season 1 matches TVDB DVD order.
Season 2 matches the TVDB aired order, which is slightly different
from the TMDB order: Aerial Assault is listed as episode 48 in TVDB
but episode 41 in TMDB. So, ... I dunno. This only adds seasons 1 and 2
because seasons 3 and 4 are even more problematic. Those are to come.