auto-conf count_episodes strategy works
I'm not sure how/if I used this strategy before, but this makes it work how I
expect it to now. I just used it for scooby doo, and it lets you specify how
many episodes are in each season, basing the season and episode solely on that.
The code isn't the prettiest, but it works.