change flintstones to transocde ac3 mono to aac mono
The main audio stream of every episode is ac3 mono. For most titles, that's also the only
audio stream. For some reason, my older fire stick + older LG tv doesn't like it when I
play this on Plex. It has no audio and seems to somehow mess up the audio generally on
the fire stick, like it doesn't play the navigation clicks afterward. A few of the titles
have an additional, stereo ac3 commentary track, which plays fine. This should change
the config to transcode just the first audio stream from ac3 to aac.