add first full configs from thediscdb
I generated these completely automatically with a set of new scripts (not yet committed)
using data from thediscdb. I manually verified each title match, and it's all good.

Note to self: I found a combined title mapped on the Rogue One bonus disc. It was a large
title that contained the content of almost all the other titles. If I encoded it as mapped,
I'd end up with duplicate content, so I removed that config file to not encode it and save
space.