new script that fully automates discdb matching
From the command mentioned a commit or two back, I built out this script, incorporating
identifying a disk in the signature db, so that I can point it at a rip dir, and it does
everything else. It worked successfully for The Meaning of Life. I had to tweak a few
things:

- File names need sanitization. They can't always be exactly the video name.
- There was a problem with requiring all discdb titles to be present in the local rip:
  titles that fall below the min-length will be missing. In The Meaning of Life, there
  was a 5-second title mapped, and it caused this to fail. I think just warning about it
  will be fine for now. This path is only meant to work if the signature db comes up with
  a 100% match, so any discrepancy should be something like this.
- The script that maps discdb types to Plex types didn't have Trailers in it yet.
- curl was being noisy