update identify-disk to work in auto-conf script
This is a combination of a few things. First, some big changes to this script have been
sitting in my working dir for months because I haven't messed with it in a while. The
big chunk of code at the end is mostly that. Second, when I came back to it, I found it
needed a bit of documentation and reorganization. That's most of the actual change here.
The last thing is that I made it write out machine-readable output, where it previously
wrote human-readable, but that was in the new big chunk of code that hadn't been committed
before anyway.