remove CUT_STREAMS from armageddon config
Armageddon main title was failng to transcode because of a stream mismatch. With
the recent advent of being able to handle all streams, including subtitles, along
with cuts, the encode script was trying to handle the additional, second audio
stream that exists in the source file. The filter script is still obeying
CUT_STREAMS, though, and it only included one audio stream in its filtergraph.
This led to encode.sh trying to map a stream that didn't exist in the filter
output. Technically, this is a bug in the interaction between the scripts, but
since I'm heading the direction of eliminating CUT_STREAMS, it's simpler to just
remove this from the config for now.