don't count encode done when only encoding one thing
This was useful in the past, when I used the encode script for everything
and sometimes pointed it at a specific file. When I pointed it at a specific
file, I wanted it to encode that, whether it was "done" or not. Now, I want
to externalize the iteration of files to the ls-inputs.sh script, and
encode.sh should always only take a single file, so its "done" logic should
work for a single file.