fix some invalid output names
These were generated by discdb matches in the early stages of figuring out how to make
that stuff work, and some invalid characters (colons) snuck into mapped file names. I
think this is fixed now by use of the sanitize function in common, so this should be
a one-time fix.