fix hardcoded base path in newtvshow script
This script had a hardcoded ripping path, which broke
when I switched drives. I decided it would just be better
to look for "/ripping/" in the path, and use that as the
base directory.

There was also some other change here that I don't recall,
which got wrapped in here.