HUGE lift to support better splitting
The first pass at trying to split, by running a new command per output, didn't work well. It
ended up spinning forever after writing the first file. That may have been because of writing
to a network mount. I'm not sure. Either way, I estimated it would take 40-50 hours to scan
one entire input file to get a single episode out of it. There are 5-6 episodes per file, so
that's insane.

I had done it that way because I thought that applying cuts to a multi-output command would be
a lot harder. Then I realized that cuts are based on inputs, so maybe it would all work out. I
went and shifted everything around to be able to build multiple outputs from the one input in
a single command, and it works! I had to do major surgery because almost all of the options I
create are output-specific and have to be applied to EVERY output.

After doing all of this, I now think, again, that cuts are going to be a problem :p

I'm pretty sure that cuts are at least totally broken right now because I didn't try to move
the code that handles subtitles via the concat filter. I also have no clue how the complex
filter is going to work in the new multi-output system or if I broke anything while moving it
around.

I may well have broken lots of things here. All I can say is that it works for a multi-episode
file of Chuck!