a little encode.sh cleanup
- Fix a major bug introduced with splitting: the done check didn't stop the script
from progressing and trying to encode. Fortunately, when "done" was detected,
it never generated the output args, so the generated command was malformed, so
it didn't mess anything up.

= Enhance the exit trap to hopefully distinguish between finished and unfinished files,
if the end of the script wasn't reached. It's only relevant to split inputs that
produce multiple outputs. It assumes an output is finished if there's a loater,
non-empty output.

Be careful! It's an EXIT trap, so it runs all the time, INCLUDING when the encode_one
function exits! That means it runs after every input file is processed successfully.
I don't think I realized this before.

- I also changed how "tv specials as movies" works slightly. I'm not sure why the
"! OUTPUTNAME =~ MAIN_NAME" check was there. For Chuck, the MAIN_NAME will almost
always be in the OUTPUTNAME. I changed it to say "if OUTPUTNAME doesn't begin with
'Season'" because tv features configured for the movie dir follow the movie naming
pattern like "Featurettes/X.mkv". I think this is a pretty good way to go.

Note: The specials for BSG worked, I think, because I set MAIN_TYPE=movie in each
title's config, not at the top level.

- Finally, I moved the "Duration" output down because it was printing even when the
encoding was already done. Unfortunately, it's not the only output that's leaking
through that check. Since I had to move the "done" check into the loop that
calculates the outputs, It's now happening a lot later than before, and the script
output is messed up as a result. I need to shuffle more things around so that
they don't happen until after the done check.