add updated validate.sh script
This is the new validate script, initially copied from movies/validate.sh, and updated for
the allineone encoder. I'm running it as

```
find allinone/data/config/ -type f -not -name main -not -name README.md -print0 | \
  xargs -0 -I {} bash -c 'result="$(allinone/validate.sh -i "{}" 2>&1)" || echo -e " -- Invalid: {}\n$result"' 2>&1 | \
  tee foo
```

Everything passes now, except House S6E1, which is split in two parts. They're mapped the
way I thought they should be, but it actually isn't working in plex, so I'm not sure what
to do with it right now.