remove dup psych mappings
All of these are mappings that are duplicated. I believe I had ripped Psych once in the old
style, nesting seasons and disks inside the top-level dir, and then plex-mageddon happened,
and I lost the ripped files. When I ripped them again, I did it in the per-disk style. I have
all the configs stored for both ways, and they show up as duplicates when I check. I verified
that all of these old-style mappings have existing, new mappings. This is mostly just for the
actual episodes. I think all of the special features might still be unmapped in the new style,
but they're all there in the old. Next step is to see if I can find a good way to match the
old, remaining mappings up to the new files.