adding an ignore
I decided to add a symlink for the encoding cache dir so that I can stop
passing CACHEDIR to the encode script every time. By default it expects
to use `cache` in the same dir as the script, but I put it on a mounted
drive instead. I added a symlink locally to that mount, but I don't want
to commit it, so ignoring it instead.