clean up validate script special title handling
I realized that the logic in validation should handle a lot of the
exceptions I'd added manually. It should just check that the OUTPUTNAME
matches the MAIN_NAME, and it should in most cases. I'm not sure why
all of those titles showed up initially. I guess I didn't have the logic
quite right yet. I can remove all of this, and only the Amazing Grace
titles show in the report, and that's right because they're weird:
two different disks put a movie file into the same directory. I'm
leaving it as a violation because I'd like to find a better way to
handle those. House s6e1 is still invalid, too, and that's also right
because it's a 2-part episode that Plex doesn't even play correctly.